Library Management System
This application is developed as part of our C# course circullum. This
application included a catalog of books which can be added and deleted.
The available books are ready to be issued and it can track the issued
books to students. Once the book is returned it readily available to
reissue.